backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
python-docx
langchain
langchain-community
sentence-transformers
chromadb
groq
motor
python-dotenv